Move quick start card related classes to cards package
